divide classes

+++ bonus-cmdr-service-ws-stub/src/main/java/kz/kcell/app/bonus_cmdr/ws/stub/StopAssigmentBonus.java
@@ -17,7 +17,7 @@ import javax.xml.bind.annotation.XmlType;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
  *         &lt;element name="cid" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/&gt;
- *         &lt;element name="bid" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/&gt;
+ *         &lt;element name="bid" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -34,7 +34,7 @@ import javax.xml.bind.annotation.XmlType;
 public class StopAssigmentBonus {
 
     protected Long cid;
-    protected Integer bid;
+    protected Long bid;
 
     /**
      * Gets the value of the cid property.
@@ -65,10 +65,10 @@ public class StopAssigmentBonus {
      * 
      * @return
      *     possible object is
-     *     {@link Integer }
+     *     {@link Long }
      *     
      */
-    public Integer getBid() {
+    public Long getBid() {
         return bid;
     }
 
@@ -77,10 +77,10 @@ public class StopAssigmentBonus {
      * 
      * @param value
      *     allowed object is
-     *     {@link Integer }
+     *     {@link Long }
      *     
      */
-    public void setBid(Integer value) {
+    public void setBid(Long value) {
         this.bid = value;
     }
